v12 Separated global operations from market controller
